refactor(api): move prompt constants to kt-agents-core, drop 2 worker deps
Addresses the architectural debt from #202 where the API imported from
three worker packages. This PR handles the prompts half of the problem:

- Move 6 LLM prompt constants from worker-synthesis and worker-nodes
  into kt-agents-core/prompts/ subpackage (inert strings, no agent logic)
- Worker modules now re-export from the shared lib for backwards compat
- Remove kt-worker-synthesis and kt-worker-nodes from API deps
- Remove their COPY lines from the API Dockerfile
- Remove dead parent resolution prompt from transparency endpoint
  (parent/pipeline.py no longer exists)

kt-worker-ingest remains as an API dependency — the prepare-stage
functions (process_ingest_sources, build_chunk_list, review_chunks) are
tightly coupled with worker internals (session rollback, emit callbacks,
fetch registry lifecycle, repo.update_fields). Extracting them cleanly
requires a dedicated kt-ingest-core lib — tracked separately.
